Further refactoring.
Now we move the extracting of tag and state to earlier in the method,
along with the associated error handling to ensure the format of the
elements is correct.

This allows us merely to test if the string starts with the count
command, and lets us reuse these extracted tags and states elsewhere.