Initial _details() implementation.
This relies on a couple of techniques to DRY out the code.  We start
by using a couple of list comprehensions to figure out which are
running and which are not.

Once we have that list we build up a response string by iterating
over the two new lists we have and constantly appending onto a base
string that contains the header.  Once we have a full string we simply
return it.