Further refactoring for details by age.
Now we extract the age, along with the other input and error check it in
the main body.  We then pass that into _details if appropriate and use
that to run a second set of list comprehensions that further winnow down
the list to only those that meet our age requirement.